rename and move optim/utils.py
